Parsing of files by jumping to substrings and counting non-commented entries
Added a routine to jump to the line containing a specific substring and a
routine to count the number of non-commented lines between two substrings.
Both routines are tested with an associated test case (test_fid_operations).
The two routines shall simplify the parsing of input files by loosening
formatting conditions.